[Rendering] Moved RenderGroup class to Rendering namespace
