+++ sources/engine/Xenko.Engine/Engine/BackgroundComponent.cs
@@ -8,6 +8,7 @@ using Xenko.Core.Annotations;
 using Xenko.Engine.Design;
 using Xenko.Engine.Processors;
 using Xenko.Graphics;
+using Xenko.Rendering;
 using Xenko.Rendering.Background;
 
 namespace Xenko.Engine
